openmetadata-mcp-server/pyproject.toml
[project]
name = "openmetadata-mcp"
version = "0.1.0"
description = "MCP Server for OpenMetadata - AI-powered data discovery and governance"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[server]>=1.1.0",
    "openmetadata-clients>=1.12.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
]
scripts = [
    "openmetadata-mcp = openmetadata_mcp.server:main",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["openmetadata_mcp"]
